refactor: 重組 `events` 檔案的路徑
