+++ src/events/_listener.ts
@@ -1,20 +1,26 @@
-import entityDie from "./after/entityDie";
-import entityHealthChange from "./after/entityHealthChange";
-import itemCompleteUse from "./after/itemCompleteUse";
-import itemUse from "./after/itemUse";
-import playerSpawn from "./after/playerSpawn";
+import entityDie from "./after/entity/entityDie";
+import entityHealthChange from "./after/entity/entityHealthChange";
+
+import itemCompleteUse from "./after/item/itemCompleteUse";
+import itemUse from "./after/item/itemUse";
+
+import playerSpawn from "./after/player/playerSpawn";
+
 import worldInit from "./after/worldInit";
+import ScriptMessage from "./after/scriptMessage";
 
 export default function eventListener() {
     const events = [
-        worldInit,
+        entityHealthChange,
+        entityDie,
 
-        playerSpawn,
         itemUse,
         itemCompleteUse,
 
-        entityHealthChange,
-        entityDie
+        playerSpawn,
+
+        worldInit,
+        ScriptMessage,
     ];
     for (const ev of events) {
         ev.subscribe();
